Worked more on page structure and also aesthetics
I attached an image of a canvas stand to the div where the user will be sketching so that it looks like an actual canvas. I also worked on the layout for the paints/erasers as well as wrote instructions on how to use the Etch-A-Sketch. Next, I will work on  more aesthetics and some responsiveness before beginning the event handling on JavaScript.